fix: handlers refactor and format
