+++ platform_plugin_turnitin/static/html/turnitin.html
@@ -19,7 +19,7 @@
             <h2>File Upload</h2>
             <form enctype="multipart/form-data">
                 <div class="file-upload">
-                    <input type="file" name="myfile" id="file" accept=".doc, .docx">
+                    <input type="file" name="uploaded_file" id="file" accept=".doc, .docx">
                     <label for="file">Select File</label>
                     <span class="selected-filename">No file selected</span>
                 </div>
