+++ setup.py
@@ -150,6 +150,9 @@ setup(
         "lms.djangoapp": [
             "platform_plugin_turnitin = platform_plugin_turnitin.apps:PlatformPluginTurnitinConfig"
         ],
+        "cms.djangoapp": [
+            "platform_plugin_turnitin = platform_plugin_turnitin.apps:PlatformPluginTurnitinConfig"
+        ],
         "xblock.v1": [
             "turnitin = platform_plugin_turnitin.turnitin:TurnitinXBlock",
         ],
